rename footer/header; show header only at path="/"
footer.jsx to Footer.jsx
header.jsx to Header.jsx
