--- src/components/header.jsx
@@ -1,13 +1,0 @@
-// import React from "react";
-
-const Header = () => {
-    return (
-        <>
-            <header className="App-header">
-                <h1>Welcome to MathMate</h1>
-            </header>
-        </>
-    )
-}
-
-export default Header
